Updating all the project - Installer successfully created

+++ src/installer/utils/Rollback.java
@@ -3,7 +3,6 @@ package installer.utils;
 import javax.swing.*;
 import java.awt.*;
 import java.io.File;
-import java.io.IOException;
 
 /**
  * Rollback class - This class will delete all the files under the directory
@@ -11,39 +10,39 @@ import java.io.IOException;
  * The GUI window will close automatically after all the files are deleted.
  */
 public class Rollback extends JDialog {
-    // Variables
-    public static final Rollback INSTANCE = new Rollback();
 
-    private Rollback(){
+    private Rollback(File directory) throws Exception{
         // Set up the window
         setUp();
+        // Delete file
+        deleteFile(directory);
     }
 
-    public void setUp(){
+    private void setUp(){
         this.setTitle("Rolling back installation...");
         this.setLocationByPlatform(true);
-        this.setVisible(true);
         this.setResizable(false);
         // Do not let the user touch any other window from this application
         this.setModal(true);
         this.setModalityType(ModalityType.APPLICATION_MODAL);
         // Do not let the user close this window
         // The window will close automatically upon completion
-        this.setDefaultCloseOperation(JDialog.DO_NOTHING_ON_CLOSE);
-        //this.setUndecorated(true);
+        this.setDefaultCloseOperation(DO_NOTHING_ON_CLOSE);
         // Set Image Icon
         this.setIconImage(InstallerUtils.createIcon("installer.png", null).getImage());
         // Set up the Progress bar
-        final JProgressBar rollbackProgressBar = new JProgressBar();
+        JProgressBar rollbackProgressBar = new JProgressBar();
         rollbackProgressBar.setPreferredSize(new Dimension(400, 15));
         rollbackProgressBar.setToolTipText("Rolling back in progress");
-        rollbackProgressBar.setStringPainted(true);
+        //rollbackProgressBar.setStringPainted(true);
         rollbackProgressBar.setIndeterminate(true);
         // Add Panels to content pane
         Container cp = this.getContentPane();
         cp.setLayout(new FlowLayout(FlowLayout.CENTER));
         cp.add(rollbackProgressBar);
         this.setContentPane(cp);
+        this.pack();
+        this.setVisible(true);
     }
 
     /**
@@ -51,13 +50,25 @@ public class Rollback extends JDialog {
      * This files will delete all the directory recursively
      * @param directory Directory to be deleted
      */
-    public void execute(File directory) throws IOException {
-        // Place a Progress bar
-        this.pack();
+    public static void execute(final File directory) {
+        SwingUtilities.invokeLater(new Runnable() {
+            @Override
+            public void run() {
+                try{
+                    new Rollback(directory);
+                } catch (Exception e){
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    private void deleteFile(File directory) throws Exception{
         this.setVisible(true);
-        // Delete the files recursively
-        InstallerUtils.deleteFile(directory);
-        // Exit the program when rollback has be completed
+        if(directory != null){
+            // Delete the files recursively
+            InstallerUtils.deleteFile(directory);
+        }
         this.setVisible(false);
     }
 }
